requirements.txt
grpcio
protobuf
Flask 
gunicorn